refactor(client): create mpeRoomUniversalSearchMachine to factorize library user mpe room & mpe room
